feat: :fire: Removed unsed files and renamed files for better clarity
